Graphs and chart added

+++ server/db/migrate/20190104095848_create_items.rb
@@ -3,8 +3,8 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.integer :user_id
       t.string :coin
-      t.integer :amount
-      t.integer :price_per_unit
+      t.float :amount
+      t.float :price_per_unit
       t.timestamps
     end
   end
